+++ src/components/Autocomplete/Autocomplete.types.ts
@@ -2,6 +2,7 @@ import type { TextFieldProps } from "components/TextField";
 import type { MenuItemProps } from "components/MenuItem";
 
 export type Props = TextFieldProps & {
+	onInput?: TextFieldProps["onChange"];
 	onItemSelect?: (args: { value: string }) => void;
 	children: React.ReactNode;
 };