+++ src/components/Container/Container.tsx
@@ -4,7 +4,17 @@ import type * as T from "./Container.types";
 import s from "./Container.module.css";
 
 const Container = (props: T.Props) => {
-	const { children, padding = 4, width, className, attributes } = props;
+	const {
+		children,
+		padding = 4,
+		width,
+		align,
+		justify,
+		height,
+		maxHeight,
+		className,
+		attributes,
+	} = props;
 	const rootClassNames = classNames(s.root, className);
 
 	return (
@@ -12,6 +22,10 @@ const Container = (props: T.Props) => {
 			attributes={attributes}
 			className={rootClassNames}
 			paddingInline={padding}
+			align={align}
+			justify={justify}
+			height={height}
+			maxHeight={maxHeight}
 			width={width}
 			maxWidth="100%"
 		>